Created a Utility for PowerShell to move the VNET of a VM.

+++ ARM/Compute/Utilities/ModifyVMVNET.ps1
@@ -1,16 +1,24 @@
-﻿
-Login-AzureRmAccount -Environment 'AzureUSGovernment'
+﻿$Environment = 'AzureCloud'
+$Environment = 'AzureUSGovernment'
+
+Login-AzureRmAccount -Environment $Environment
 $subscription = Get-AzureRmSubscription |  Out-GridView -PassThru
 Set-AzureRmContext -SubscriptionId $subscription.Id
 Write-Host "Successfully logged in to Azure." -ForegroundColor Green 
-$VMS = Get-AzureRmVM 
+
+
+
+#Limit the VNET migration to a specific Resource Group.
+$ResouceGroup = Get-AzureRMResourceGroup | Out-GridView -PassThru
+$VMs= Get-AzureRMVM -ResourceGroupName $ResouceGroup.ResourceGroupName
+#Limit the VM migration to a specific VM and just bipass the ForEach below.
+#$VM = Get-AzureRMVM -ResourceGroupName $ResouceGroup.ResourceGroupName | Out-GridView -PassThru
 
 Foreach($VM in $VMS)
 {
-    if($Vm[0].AvailabilitySetReference){}
-
+    write-output $vm.name + $vm.Location
 #Get the current VM's NIC name to get the NIC variable set so the picklist of Subnets will allow the user to have an idea where the NIC was
-
+<# #> 
     $NicName = (($VM.NetworkProfile.NetworkInterfaces[0].id).Split('/'))[8]
     $Nic = Get-AzureRmNetworkInterface -Name $NicName -ResourceGroupName $VM.ResourceGroupName
     $Title = (($NIC[0].IpConfigurations.subnet.id).split('/'))[10]
@@ -22,32 +30,63 @@ Foreach($VM in $VMS)
     $NewSubnet = Get-AzureRmVirtualNetworkSubnetConfig -VirtualNetwork $NewNetwork -Name $NewSubnet.Name
     
     $message = "current name: " + $vm.name + ", Current IP: " +$NIC[0].IpConfigurations.privateIPAddress+ ", New IP must be in this prefex: " + $NewSubnet.AddressPrefix
-
-
     $NewPrivateIP = Read-Host $message
+    $NewIPConfig = New-AzureRmNetworkInterfaceIpConfig -Subnet $NewSubnet -Name 'config1' -Primary -PrivateIpAddress $NewPrivateIP 
 
-    $NewIPConfig = New-AzureRmNetworkInterfaceIpConfig -SubnetId $NewSubnet.Id -Name 'config1' -Primary -PrivateIpAddress $NewPrivateIP -PublicIpAddressId 
-
-    $NewNic = New-AzureRmNetworkInterface -Name "$($NicName)1a" -ResourceGroupName $Nic.ResourceGroupName `
-                                          -Location $Nic.Location -IpConfiguration $NewIPConfig -Force
+    
+    $NewNic = New-AzureRmNetworkInterface -Name "$($NicName)1a" -ResourceGroupName $Nic.ResourceGroupName -Location $Nic.Location -IpConfiguration $NewIPConfig -Force 
    
 
    #using the VM Configuration in memory, remove the primary NIC and replace with the new NIC
-    Remove-AzureRmVMNetworkInterface -VM $VM -NetworkInterfaceIDs $NewNic.Id
+    $VM = Remove-AzureRmVMNetworkInterface -VM $VM -NetworkInterfaceIDs $Nic.Id
 
-    Add-AzureRmVMNetworkInterface -VM $VM -NetworkInterface $NewNic
+    $vm = Add-AzureRmVMNetworkInterface -VM $VM -NetworkInterface $NewNic
+    $Proceed = Read-host 'Type Yes if you want to proceed'
+    if($Proceed -eq "Yes"){
+        Stop-AzureRmVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName -Force
+        Remove-AzureRmVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName -Force
 
-    Stop-AzureRmVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName -Force
-    Remove-AzureRmVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName -Force
-     New-AzureRmVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName
+        #Is there an Availability Set?
+        if($vm.AvailabilitySetReference.id -eq $null){
+            $newVMConfig = new-azurermvmconfig -VMSize $vm.HardwareProfile.VmSize -VMName $vm.Name -Tags $vm.tags 
+        }
+        else{
+        $newVMConfig = new-azurermvmconfig -VMSize $vm.HardwareProfile.VmSize -VMName $vm.Name -AvailabilitySetId $vm.AvailabilitySetReference.Id -Tags $vm.tags 
+        }
+     
+        Add-AzureRmVMNetworkInterface -vm $newVMConfig -id $vm.NetworkProfile.NetworkInterfaces[0].Id
+        #Is the VM Windows or Linux?
+        if($vm.StorageProfile.osDisk.OStype -eq "Windows"){
+            #Is the VM using Managed Disk or Storage Accounts?
+            if($vm.storageProfile.OSDisk.ManagedDisk.ID -eq $null){
+                Set-AzureRmVMOSDisk -vm $newVMConfig -Name $vm.StorageProfile.OsDisk.Name -CreateOption Attach -vhdURI $vm.StorageProfile.OsDisk.Vhd.Uri -Windows 
+            }
+            else{
+                Set-AzureRmVMOSDisk -vm $newVMConfig -Name $vm.StorageProfile.OsDisk.Name -CreateOption Attach -ManagedDiskId $vm.StorageProfile.OsDisk.ManagedDisk.Id -Windows 
+            }
+        }
+        else {
+            Set-AzureRmVMOSDisk -vm $newVMConfig -Name $vm.StorageProfile.OsDisk.Name -CreateOption Attach -ManagedDiskId $vm.StorageProfile.OsDisk.ManagedDisk.Id -Linux
+        }
+        #Does the VM have any Data Disks?
+        if($vm.StorageProfile.DataDisks -ne $null){
+            foreach($Disk in $VM.StorageProfile.DataDisks){
+                if($Disk.ManagedDisk -eq $null){
+                    add-AzureRmVMDataDisk -Name $disk.name -Lun $disk.lun -VM $newVMConfig -Caching $disk.caching -CreateOption attach -VhdUri $disk.vhd.Uri
+                }
+                else{
+                    add-AzureRmVMDataDisk -Name $disk.name -Lun $disk.lun -VM $newVMConfig -Caching $disk.caching -CreateOption attach -ManagedDiskId $disk.ManagedDisk
 
-       #   New-AzureRmVM -Name $VM.Name -ResourceGroupName 'rg-ActiveDirectoryDomainServices-tx'
-       #    Stop-AzureRmVM -Name $VM.Name -ResourceGroupName 'rg-ActiveDirectoryDomainServices-tx' -Force
-    #Remove-AzureRmVM -Name $VM.Name -ResourceGroupName 'rg-ActiveDirectoryDomainServices-tx' -Force
- #$newVM = Get-azurermVM -Name $VM.Name -ResourceGroupName 'rg-ActiveDirectoryDomainServices-tx'
-    Start-AzureRmVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName
-  $newVM = Get-azurermVM -Name $VM.Name -ResourceGroupName $Vm.ResourceGroupName
-   $newVM.NetworkProfile.NetworkInterfaces[0].Id
+                }
+            }
+        }
+
+
+
+    $NewVM = $newVMConfig | New-AzureRmVM -ResourceGroupName $vm.ResourceGroupName -Location $vm.Location
+
+
+    }
+
 }
-
